CommonComReader: Only allow conversion of one file
Otherwise we are mixing up files with each other.
It is not only safer to do it like that generally, it is also confirmed
by Dessault Systems that it is only possible to talk with one SolidWorks
instance at the end of the COM connection, where the ROT is managing the
service.